+++ app/src/main.c
@@ -18,7 +18,7 @@
 #include "util/net.h"
 #include "util/term.h"
 #include "version.h"
-#include "microphone.h"
+#include "client_audio.h"
 
 #ifdef _WIN32
 #include <windows.h>
